Refactor submission JSX to modular setup
- Move things into modules by 'app'
- Use ES6 modules' "index.js" feature
- Use extension JSX only for components
