--- src/resources/assets/js/components/Submission/Pages.jsx
@@ -1,10 +1,0 @@
-import Scoreboard from './Scoreboard.jsx'
-import ClassList from './ClassList.jsx'
-import Courses from './Courses.jsx'
-import Review from './Review.jsx'
-
-export {Scoreboard, ClassList, Courses, Review};
-
-
-export { ApplicationsIndex, ApplicationsEdit } from './Applications'
-
