+++ src/resources/assets/js/main.jsx
@@ -7,12 +7,12 @@ import { syncHistoryWithStore, routerReducer } from 'react-router-redux'
 import thunk from 'redux-thunk'
 
 import { LiveScoreboard } from './live_scoreboard'
-import SubmissionFlow from './components/SubmissionFlow'
-import { submissionReducer } from './states/SubmissionStates'
+import { SubmissionFlow } from './submission'
+import { submissionReducer } from './submission/reducers'
 
 const reducer = combineReducers({
-	routing: routerReducer,
-	submission: submissionReducer
+    routing: routerReducer,
+    submission: submissionReducer
 })
 
 
@@ -21,24 +21,24 @@ const store = createStore(reducer, undefined, compose(applyMiddleware(thunk), wi
 const history = syncHistoryWithStore(browserHistory, store)
 
 function _wrapProvider(v) {
-	return (
-		<Provider store={store}>
-			<Router history={history}>
-				{v}
-			</Router>
-		</Provider>
-	);
+    return (
+        <Provider store={store}>
+            <Router history={history}>
+                {v}
+            </Router>
+        </Provider>
+    );
 }
 
 var _components = [
-	['#live-scoreboard', function(elem) { ReactDOM.render(<LiveScoreboard/>, elem) }],
-	['#submission-flow', function(elem) { ReactDOM.render(_wrapProvider(SubmissionFlow()), elem); }]
+    ['#live-scoreboard', function(elem) { ReactDOM.render(<LiveScoreboard/>, elem) }],
+    ['#submission-flow', function(elem) { ReactDOM.render(_wrapProvider(SubmissionFlow()), elem); }]
 ]
 
 
 _components.forEach(function(c) {
-	var elem = document.querySelector(c[0]);
-	if (elem) {
-		c[1](elem);
-	}
+    var elem = document.querySelector(c[0]);
+    if (elem) {
+        c[1](elem);
+    }
 });
