+++ src/resources/assets/js/submission/Scoreboard.jsx
