+++ src/resources/assets/js/submission/applications/index.js
@@ -1,0 +1,1 @@
+export { ApplicationsIndex, ApplicationsEdit } from './components'
