+++ src/resources/assets/js/submission/base_components.jsx
