+++ src/resources/assets/js/submission/class_list/index.js
@@ -1,0 +1,3 @@
+import ClassList from './ClassList'
+
+export { ClassList }
