+++ src/resources/assets/js/submission/index.js
@@ -1,0 +1,3 @@
+import SubmissionFlow from './SubmissionFlow'
+
+export { SubmissionFlow }
