+++ src/resources/assets/js/submission/pages.js
@@ -1,0 +1,9 @@
+import Scoreboard from './Scoreboard'
+import Courses from './Courses'
+import Review from './Review'
+
+export {Scoreboard, Courses, Review}
+
+
+export { ApplicationsIndex, ApplicationsEdit } from './applications'
+export { ClassList } from './class_list'
